Working on dummy data insert
Dummy data insert is still not working.  I plan on creating tab
delimited files in order to insert the data, but I need to get the query
right.

The Install button in the install.php page creates/resets the database
successfully.

I modified a few comments in your column and index files, but everything
looks good for those so I moved the card to done.